move all to flush, we don't need a separate package
